Extract service for extracting the template name
